+++ packages/orq-rc/RELEASES.md
@@ -5249,3 +5249,13 @@ Based on:
 - [typescript v4.0.0-rc.59] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.59 - packages/orq-rc
+
+## 2025-11-20 15:30:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.60] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.60 - packages/orq-rc